Move check constraint violation logic to builder
