+++ common/persistence/src/test/java/app/coronawarn/server/common/persistence/domain/DiagnosisKeyBuilderTest.java
@@ -28,6 +28,9 @@ import app.coronawarn.server.common.protocols.external.exposurenotification.Key;
 import com.google.protobuf.ByteString;
 import java.nio.charset.Charset;
 import java.time.Instant;
+import java.time.LocalDate;
+import java.time.ZoneOffset;
+import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 
 public class DiagnosisKeyBuilderTest {
@@ -143,6 +146,119 @@ public class DiagnosisKeyBuilderTest {
     );
   }
 
+  @Test
+  public void transmissionRiskLevelMustBeInRange() {
+    checkMessage(assertThrows(
+        InvalidDiagnosisKeyException.class, () -> keyWithRiskLevel(9)),
+        "[Risk level must be between 0 and 8. Invalid Value: 9]");
+
+    checkMessage(assertThrows(
+        InvalidDiagnosisKeyException.class, () -> keyWithRiskLevel(-1)),
+        "[Risk level must be between 0 and 8. Invalid Value: -1]");
+  }
+
+  @Test
+  public void transmissionRiskLevelDoesNotThrowForValid() {
+    Assertions.assertThatCode(() -> keyWithRiskLevel(0)).doesNotThrowAnyException();
+    Assertions.assertThatCode(() -> keyWithRiskLevel(8)).doesNotThrowAnyException();
+  }
+
+  @Test
+  public void rollingStartNumberDoesNotThrowForValid() {
+    Assertions.assertThatCode(() -> keyWithRollingStartNumber(4200L)).doesNotThrowAnyException();
+
+    // Timestamp: 05/16/2020 @ 00:00 in hours
+    Assertions.assertThatCode(() -> keyWithRollingStartNumber(441552L)).doesNotThrowAnyException();
+  }
+
+  @Test
+  public void rollingStartNumberCannotBeInFuture() {
+    checkMessage(assertThrows(
+        InvalidDiagnosisKeyException.class,
+        () -> keyWithRollingStartNumber(1904169600L)),
+        "[Rolling start number must be greater 0 and cannot be in the future. Invalid Value: 1904169600]");
+
+    long tomorrow = LocalDate
+        .ofInstant(Instant.now(), ZoneOffset.UTC)
+        .plusDays(1).atStartOfDay()
+        .toEpochSecond(ZoneOffset.UTC);
+
+    checkMessage(assertThrows(
+        InvalidDiagnosisKeyException.class, () -> keyWithRollingStartNumber(tomorrow)),
+        String.format("[Rolling start number must be greater 0 and cannot be in the future. Invalid Value: %s]",
+            tomorrow));
+
+  }
+
+  @Test
+  public void rollingPeriodMustBeLargerThanZero() {
+    checkMessage(assertThrows(
+        InvalidDiagnosisKeyException.class, () -> keyWithRollingPeriod(0)),
+        "[Rolling period must be greater than 0. Invalid Value: 0]");
+
+    checkMessage(assertThrows(
+        InvalidDiagnosisKeyException.class, () -> keyWithRollingPeriod(-3L)),
+        "[Rolling period must be greater than 0. Invalid Value: -3]");
+  }
+
+  @Test
+  public void rollingPeriodDoesNotThrowForValid() {
+    Assertions.assertThatCode(() -> keyWithRollingPeriod(144L)).doesNotThrowAnyException();
+  }
+
+  @Test
+  public void keyDataMustHaveValidLength() {
+    assertThrows(
+        InvalidDiagnosisKeyException.class,
+        () -> keyWithKeyData("17--bytelongarray".getBytes(Charset.defaultCharset())));
+
+    assertThrows(
+        InvalidDiagnosisKeyException.class,
+        () -> keyWithKeyData("".getBytes(Charset.defaultCharset())));
+
+    assertThrows(
+        InvalidDiagnosisKeyException.class,
+        () -> keyWithKeyData("1".getBytes(Charset.defaultCharset())));
+  }
+
+  @Test
+  public void keyDataDoesNotThrowOnValid() {
+    Assertions.assertThatCode(() -> keyWithKeyData("16-bytelongarray".getBytes(Charset.defaultCharset())))
+        .doesNotThrowAnyException();
+  }
+
+  private DiagnosisKey keyWithKeyData(byte[] expKeyData) {
+    return DiagnosisKey.builder()
+        .withKeyData(expKeyData)
+        .withRollingStartNumber(expRollingStartNumber)
+        .withRollingPeriod(expRollingPeriod)
+        .withTransmissionRiskLevel(expTransmissionRiskLevel).build();
+  }
+
+  private DiagnosisKey keyWithRollingStartNumber(long expRollingStartNumber) {
+    return DiagnosisKey.builder()
+        .withKeyData(expKeyData)
+        .withRollingStartNumber(expRollingStartNumber)
+        .withRollingPeriod(expRollingPeriod)
+        .withTransmissionRiskLevel(expTransmissionRiskLevel).build();
+  }
+
+  private DiagnosisKey keyWithRollingPeriod(long expRollingPeriod) {
+    return DiagnosisKey.builder()
+        .withKeyData(expKeyData)
+        .withRollingStartNumber(expRollingStartNumber)
+        .withRollingPeriod(expRollingPeriod)
+        .withTransmissionRiskLevel(expTransmissionRiskLevel).build();
+  }
+
+  private DiagnosisKey keyWithRiskLevel(int expTransmissionRiskLevel) {
+    return DiagnosisKey.builder()
+        .withKeyData(expKeyData)
+        .withRollingStartNumber(expRollingStartNumber)
+        .withRollingPeriod(expRollingPeriod)
+        .withTransmissionRiskLevel(expTransmissionRiskLevel).build();
+  }
+
   private void assertDiagnosisKeyEquals(DiagnosisKey actDiagnosisKey) {
     assertDiagnosisKeyEquals(actDiagnosisKey, getCurrentHoursSinceEpoch());
   }
@@ -157,6 +273,10 @@ public class DiagnosisKeyBuilderTest {
     assertEquals(this.expRollingStartNumber, actDiagnosisKey.getRollingStartNumber());
     assertEquals(this.expRollingPeriod, actDiagnosisKey.getRollingPeriod());
     assertEquals(this.expTransmissionRiskLevel, actDiagnosisKey.getTransmissionRiskLevel());
+  }
+
+  private void checkMessage(InvalidDiagnosisKeyException ex, String message) {
+    assertEquals(ex.getMessage(), message);
   }
 
 }
